+++ src/views/domain-schedules/domain-schedules-create-modal/create-schedule-form/create-schedule-form.types.ts
