+++ src/views/domain-schedules/domain-schedules-create-modal/domain-schedules-create-modal.types.ts
