+++ src/views/domain-schedules/domain-schedules-create-modal/schemas/create-schedule-form-schema.ts
