+++ src/views/domain-schedules/domain-schedules-horizontal-field/domain-schedules-horizontal-field.styles.ts
